Apply suggestions from code review

+++ src/abstractdataframe/join.jl
@@ -122,9 +122,9 @@ function compose_joined_table(joiner::DataFrameJoiner, kind::Symbol,
         # assign 0x1 to it anyway and these rows are guaranteed to come first
         # (even if they are permuted)
         left_indicator = zeros(UInt32, nrow)
-        left_indicator[axes(all_orig_left_ixs, 1)] .= UInt32(1)
+        left_indicator[axes(all_orig_left_ixs, 1)] .= 1
         right_indicator = zeros(UInt32, nrow)
-        right_indicator[axes(all_orig_right_ixs, 1)] .= UInt32(2)
+        right_indicator[axes(all_orig_right_ixs, 1)] .= 2
         permute!(right_indicator, right_perm)
     end
 
@@ -418,9 +418,9 @@ function _join(df1::AbstractDataFrame, df2::AbstractDataFrame;
     if indicator !== nothing
         left_indicator .+= right_indicator
         pool = ["left_only", "right_only", "both"]
-        invpool = Dict{String, UInt8}("left_only" => UInt32(1),
-                                      "right_only" => UInt32(2),
-                                      "both" => UInt32(3))
+        invpool = Dict{String, UInt32}("left_only" => 1,
+                                       "right_only" => 2,
+                                       "both" => 3)
         indicatorcol = PooledArray(PooledArrays.RefArray(left_indicator),
                                    invpool, pool)
